+++ src/frontend/apps/desk/src/features/mail-domains/access-management/api/index.ts
@@ -1,3 +1,5 @@
 export * from './useMailDomainAccesses';
 export * from './useUpdateMailDomainAccess';
+export * from './useCreateMailDomainAccess';
 export * from './useDeleteMailDomainAccess';
+export * from './useCreateInvitation';